refactor: extract main controller to component
Upgrade to HSL 2.x - preparation
